% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.campbell}
\alias{process.campbell}
\title{Process CAMPBELL SCIENTIFIC data.}
\usage{
process.campbell(aws_dir, adt_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder on CAMPBELL server.}

\item{adt_dir}{full path to the directory containing the AWS_DATA folder on ADT server.}
}
\description{
Get the data from CAMPBELL SCIENTIFIC database, parse and convert into ADT format.
}
